refactor(): remove hpke unused files
