+++ src/utils/stampers/bundleStamper/hpke/cipherSuiteNative.ts
@@ -49,70 +49,6 @@ const LABEL_SECRET = new Uint8Array([115, 101, 99, 114, 101, 116]);
 // deno-fmt-ignore
 const SUITE_ID_HEADER_HPKE = new Uint8Array([72, 80, 75, 69, 0, 0, 0, 0, 0, 0]);
 
-/**
- * The Hybrid Public Key Encryption (HPKE) ciphersuite,
- * which is implemented using only
- * {@link https://www.w3.org/TR/WebCryptoAPI/ | Web Cryptography API}.
- *
- * This is the super class of {@link CipherSuite} and the same as
- * {@link https://deno.land/x/hpke/core/mod.ts?s=CipherSuite | @hpke/core#CipherSuite },
- * which supports only the ciphersuites that can be implemented on the native
- * {@link https://www.w3.org/TR/WebCryptoAPI/ | Web Cryptography API}.
- * Therefore, the following cryptographic algorithms are not supported for now:
- *   - DHKEM(X25519, HKDF-SHA256)
- *   - DHKEM(X448, HKDF-SHA512)
- *   - ChaCha20Poly1305
- *
- * In addtion, the HKDF functions contained in this class can only derive
- * keys of the same length as the `hashSize`.
- *
- * If you want to use the unsupported cryptographic algorithms
- * above or derive keys longer than the `hashSize`,
- * please use {@link CipherSuite}.
- *
- * This class provides following functions:
- *
- * - Creates encryption contexts both for senders and recipients.
- *   - {@link createSenderContext}
- *   - {@link createRecipientContext}
- * - Provides single-shot encryption API.
- *   - {@link seal}
- *   - {@link open}
- *
- * The calling of the constructor of this class is the starting
- * point for HPKE operations for both senders and recipients.
- *
- * @example Use only ciphersuites supported by Web Cryptography API.
- *
- * ```ts
- * import {
- *   Aes128Gcm,
- *   DhkemP256HkdfSha256,
- *   HkdfSha256,
- *   CipherSuite,
- * } from "http://deno.land/x/hpke/mod";
- *
- * const suite = new CipherSuite({
- *   kem: new DhkemP256HkdfSha256(),
- *   kdf: new HkdfSha256(),
- *   aead: new Aes128Gcm(),
- * });
- * ```
- *
- * @example Use a ciphersuite which is currently not supported by Web Cryptography API.
- *
- * ```ts
- * import { Aes128Gcm, HkdfSha256, CipherSuite } from "http://deno.land/x/hpke/mod";
- * // Use an extension module.
- * import { DhkemX25519HkdfSha256 } from "https://deno.land/x/hpke/x/dhkem-x25519/mod";
- *
- * const suite = new CipherSuite({
- *   kem: new DhkemX25519HkdfSha256(),
- *   kdf: new HkdfSha256(),
- *   aead: new Aes128Gcm(),
- * });
- * ```
- */
 export class CipherSuiteNative extends NativeAlgorithm {
   protected _kem: KemInterface;
   private _kdf: KdfInterface;
