--- src/utils/stampers/bundleStamper/hpke/interfaces/dhkemInterface.ts
@@ -1,9 +1,0 @@
-import type { KdfInterface } from './kdfInterface';
-import type { KemInterface } from './kemInterface';
-
-/**
- * The DHKEM interface.
- */
-export interface DhkemInterface extends KemInterface {
-  readonly kdf: KdfInterface;
-}
