+++ src/utils/stampers/bundleStamper/hpke/interfaces/dhkemPrimitives.ts
@@ -1,11 +1,6 @@
-// The key usages for KEM.
-export const KEM_USAGES: KeyUsage[] = ['deriveBits'];
-
-// b"dkp_prk"
 export const LABEL_DKP_PRK = new Uint8Array([100, 107, 112, 95, 112, 114, 107]);
 
-// b"sk"
-export const LABEL_SK = new Uint8Array([115, 107]);
+export const KEM_USAGES: KeyUsage[] = ['deriveBits'];
 
 export interface DhkemPrimitives {
   serializePublicKey(key: CryptoKey): Promise<ArrayBuffer>;
