--- src/utils/stampers/bundleStamper/hpke/kdfs/hkdfSha384.ts
@@ -1,31 +1,0 @@
-/* eslint-disable */
-
-// @ts-ignore: for "npm:"
-import { hmac } from 'npm:@noble/hashes@1.3.3/hmac';
-// @ts-ignore: for "npm:"
-import { sha384 } from 'npm:@noble/hashes@1.3.3/sha512';
-
-import { HkdfSha384Native } from './hkdf';
-
-export class HkdfSha384 extends HkdfSha384Native {
-  public override async extract(
-    salt: ArrayBuffer,
-    ikm: ArrayBuffer,
-  ): Promise<ArrayBuffer> {
-    await this._setup();
-    if (salt.byteLength === 0) {
-      salt = new ArrayBuffer(this.hashSize);
-    }
-    if (salt.byteLength !== this.hashSize) {
-      return hmac(sha384, new Uint8Array(salt), new Uint8Array(ikm));
-    }
-    const key = await (this._api as SubtleCrypto).importKey(
-      'raw',
-      salt,
-      this.algHash,
-      false,
-      ['sign'],
-    );
-    return await (this._api as SubtleCrypto).sign('HMAC', key, ikm);
-  }
-}
