+++ src/utils/stampers/bundleStamper/hpke/kems/dhkemNative.ts
@@ -1,9 +1,5 @@
 import { KemId } from '../identifiers';
-import {
-  HkdfSha256Native,
-  HkdfSha384Native,
-  HkdfSha512Native,
-} from '../kdfs/hkdf';
+import { HkdfSha256Native } from '../kdfs/hkdf';
 import { Dhkem } from './dhkem';
 import { Ec } from './dhkemPrimitives/ec';
 
@@ -18,33 +14,5 @@ export class DhkemP256HkdfSha256Native extends Dhkem {
     const kdf = new HkdfSha256Native();
     const prim = new Ec(KemId.DhkemP256HkdfSha256, kdf);
     super(KemId.DhkemP256HkdfSha256, prim, kdf);
-  }
-}
-
-export class DhkemP384HkdfSha384Native extends Dhkem {
-  public readonly id: KemId = KemId.DhkemP384HkdfSha384;
-  public readonly secretSize: number = 48;
-  public readonly encSize: number = 97;
-  public readonly publicKeySize: number = 97;
-  public readonly privateKeySize: number = 48;
-
-  constructor() {
-    const kdf = new HkdfSha384Native();
-    const prim = new Ec(KemId.DhkemP384HkdfSha384, kdf);
-    super(KemId.DhkemP384HkdfSha384, prim, kdf);
-  }
-}
-
-export class DhkemP521HkdfSha512Native extends Dhkem {
-  public readonly id: KemId = KemId.DhkemP521HkdfSha512;
-  public readonly secretSize: number = 64;
-  public readonly encSize: number = 133;
-  public readonly publicKeySize: number = 133;
-  public readonly privateKeySize: number = 64;
-
-  constructor() {
-    const kdf = new HkdfSha512Native();
-    const prim = new Ec(KemId.DhkemP521HkdfSha512, kdf);
-    super(KemId.DhkemP521HkdfSha512, prim, kdf);
   }
 }
