--- src/utils/stampers/bundleStamper/hpke/kems/primitives/kyber/errors.ts
@@ -1,6 +1,0 @@
-import { BaseError } from '../../../errors';
-
-/**
- * The base error class of kyber.
- */
-export class KyberError extends BaseError {}
